refactor(replay): split target verification into engine policy and daemon authority (#1478 P5)
